391: create look/playerLocations.js for findEnemyFC() and playerIsNearPoint()
